Add User and Task reation using foreign key

+++ app/routes/tasks.py
@@ -1,14 +1,14 @@
 from flask import Blueprint, render_template, request, redirect, url_for, flash, session
 from app import db
 from app.models import Task
-from flask_login import login_required
+from flask_login import login_required, current_user
 
 tasks_bp = Blueprint('tasks', __name__)
 
 @tasks_bp.route('/')
 @login_required
 def view_tasks():
-    tasks = Task.query.all()
+    tasks = Task.query.filter_by(user_id=current_user.id).all()
     return render_template('tasks.html',tasks=tasks)
 
 @tasks_bp.route('/add',methods=["POST"])
@@ -16,7 +16,7 @@ def view_tasks():
 def add_tasks():
    title = request.form.get('title')
    if title:
-       new_task = Task(title=title, status="Pending")
+       new_task = Task(title=title, status="Pending", user_id=current_user.id)
        db.session.add(new_task)
        db.session.commit()
        flash('Task Added Successfully','success')
@@ -37,10 +37,10 @@ def toggle_status(task_id):
         db.session.commit()
     return redirect(url_for('tasks.view_tasks'))
 
-@tasks_bp.route('/clear',methods=['POST'])
+@tasks_bp.route('/clear/',methods=['POST'])
 @login_required
 def clear_task():
-    Task.query.delete()
+    Task.query.filter_by(user_id=current_user.id).delete()
     db.session.commit()
     flash('Clear All Tasks','success')
     return redirect(url_for('tasks.view_tasks'))
